requirements.txt
asgiref==3.3.1
astroid==2.4.2
backcall==0.2.0
colorama==0.4.4
decorator==4.4.2
Django==3.1.3
django-crispy-forms==1.10.0
djangorestframework==3.12.2
gunicorn==20.0.4
joblib==0.17.0
numpy==1.19.4
pandas==1.1.4
parso==0.7.1
pickleshare==0.7.5
prompt-toolkit==3.0.8
psycopg2==2.8.6
Pygments==2.7.2
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.4
pywin32==300
pyzmq==20.0.0
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
wrapt==1.12.1
